requirements.txt
requests
tqdm
gunicorn
flask==3.1.1
azure-storage-blob==12.26.0
azure-batch==14.2.0
azure-cosmos==4.7.0
megadetector==10.0.3